Rename MTN Customer Churn rate.pbix to MTN Customer Churn Rate.pbix
This dataset was analyzed using Power BI after downloading the raw data from Kaggle website. It explicitly shows the churn rate of MTN customers in the first quarter of 2025, the reasons for the churn, the average income spent by customers on MTN products in that period, the average and total revenue generated by MTN during that period and a whole lot more relevant information.